Move view matrix mutiplication to fragment shader
This is slower but it sets the stage for higher precision emulation in
the fragment shader. We can't control the float precision that the gpu
uses for interpolation, so we must make sure it's interpolating on a
range where it has enough precision. i.e. not across a highly zoomed in
portion of the complex plane.